+++ unity/Assets/Modules/Calibration/Scripts/MultiMarker_MarkerSetup.cs
@@ -53,7 +53,7 @@ namespace Assets.Modules.Calibration
 
             _markerPreview = Instantiate(MarkerPreviewPrefab);
 
-            var markerSize = (float)ArucoListener.Instance.MarkerSizeInMeter;
+            var markerSize = ArucoListener.Instance.MarkerSizeInMeter;
             _markerPreview.transform.localScale = new Vector3(markerSize, 0.001f, markerSize);
 
             OptitrackListener.Instance.PosesReceived += OnOptitrackPoses;
