+++ unity/Assets/Modules/InteractiveSurface/Scripts/TableSetup_Aruco.cs
@@ -125,7 +125,7 @@ namespace Assets.Modules.InteractiveSurface
             // invert rotation on diagonal so that forward == 0,0,1
             diagonal = Quaternion.Inverse(avgRotation) * diagonal;
 
-            var markerSize = (float)ArucoListener.Instance.MarkerSizeInMeter;
+            var markerSize = ArucoListener.Instance.MarkerSizeInMeter;
             // add borderWidth + half marker size (since position is from marker center) once for both sides (*2)
             var scale = new Vector3(Mathf.Abs(diagonal.x) + 2 * BorderWidthCm / 100 + markerSize, 0.05f, Mathf.Abs(diagonal.z) + 2 * BorderWidthCm / 100 + markerSize);
